requirements.txt
Django==2.2.7
django-filter==2.2.0
djangorestframework==3.10.3
gevent==1.5a2
gunicorn==20.0.0
pytest==5.2.4
pytest-cov==2.8.1
pytest-django==3.7.0
psycopg2==2.8.4
passlib==1.7.1